requirement.txt
aniso8601==6.0.0
asn1crypto==0.24.0
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
Flask==1.0.2
Flask-Dance==1.4.0
Flask-PyMongo==2.2.0
flask-restplus==0.12.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
jwt==0.6.1
lazy==1.4
MarkupSafe==1.1.1
oauthlib==3.0.1
pycparser==2.19
pymodm==0.4.1
pymongo==3.7.2
pyrsistent==0.14.11
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.1
URLObject==2.4.3
Werkzeug==0.14.1